requirements.txt
Flask
fastapi
openai
openai-whisper
moviepy
transformers
uvicorn
gunicorn